FINISH LINE: Mostly working up through CountdownScreen

+++ libraries/scooter_base/LcdComponent.cpp
@@ -6,7 +6,7 @@
 
 extern StateMachine* g_stateMachine;
 
-void LcdComponent::handleMessage(uint8_t cmd, uint8_t len, byte* buffer) {
+void LcdComponent::handleMessage(uint16_t cmd, uint8_t len, byte* buffer) {
   _ready = true;
   g_stateMachine->handleMessage(cmd, len, buffer);
 }
